refactor: logo svg to image
