+++ apps/web/src/components/common/logo/index.ts
